Add files via upload
In this project, I have gone through the complete process of building a machine learning model based on the Loan Prediction dataset. Where I  have used different information available of different applicants regarding their background such as whether they are Married or not, about their Sex, Education level, Income of applicant, Loan amount, Property area to be used as mortgage, Loan amount etc. So in this problem, we have to predict where an applicant will receive a loan or not.
Machine learning models 
     Logistic Regression
    Decisison Tree Model
    Random Forest
    Neural Network
    Gradient Boosting
    K-Nearest Neighbors
    Support Vector Machine Classifier
    Naive Bayes